Liaison de la partie GUI et logique du jeu de la vie

+++ GridNormal.cpp
@@ -55,17 +55,17 @@ void GridNormal::updateCell(int x, int y, Rules* rules) {
 }
 Grid* GridNormal::clone() {
     vector<vector<CellState*>> newCells(height);
-    for (size_t y = 0; y < height; y++) {
+    for (int y = 0; y < height; y++) {
         newCells[y].resize(width);
-        for (size_t x = 0; x < width; x++) {
+        for (int x = 0; x < width; x++) {
             newCells[y][x] = cells[y][x]->clone();
         }
     }
     return new GridNormal(width, height, newCells);
 }
 void GridNormal::clear() {
-    for (size_t y = 0; y < height; y++) {
-        for (size_t x = 0; x < width; x++) {
+    for (int y = 0; y < height; y++) {
+        for (int x = 0; x < width; x++) {
             delete cells[y][x];
         }
     }
